barchart: rename StackedBarChart with BarChart
