+++ src/lib/components/barchart/BarChart.component.test.js
@@ -1,5 +1,5 @@
 import initStoryshots from "@storybook/addon-storyshots";
 
 initStoryshots({
-  configPath: "src/lib/components/stackedbarchart/__snapshots__"
+  configPath: "src/lib/components/barchart/__snapshots__"
 });
