+++ src/lib/index.js
@@ -29,7 +29,7 @@ import ProgressBar from "./components/progressbar/ProgressBar.component";
 import TextArea from "./components/textarea/TextArea.component";
 import CloudProgressBar from "./components/cloudprogressbar/CloudProgressBar.component";
 import Sparkline from "./components/sparkline/SparkLine.component.js";
-import StackedBarChart from "./components/stackedbarchart/StackedBarChart.component";
+import BarChart from "./components/barchart/BarChart.component";
 
 export {
   LOADER_SIZE,
@@ -60,5 +60,5 @@ export {
   CloudProgressBar,
   Sparkline,
   Banner,
-  StackedBarChart
+  BarChart
 };
