+++ stories/index.js
@@ -27,4 +27,4 @@ import "./banner";
 import "./collapsiblepanel";
 import "./cloudprogressbar";
 import "./sparkline";
-import "./stackedbarchart";
+import "./barchart";
